serde MerkleHash into hex string
